requirements.txt
# Core
python-dotenv>=1.0.0
gradio>=4.0.0

# LangChain & Ecosystem
langchain>=0.2.0
langchain-community>=0.2.0
langchain-huggingface>=0.0.3
langchain-core>=0.2.0
langchain-groq>=0.0.1

# Vector DB
faiss-cpu>=1.7.4

# HuggingFace Embeddings
sentence-transformers>=2.7.0
transformers>=4.44.0
torch>=2.2.0
accelerate>=0.33.0

# Image Handling
Pillow>=10.2.0

# Groq SDK
groq>=0.9.0

# Utilities
tqdm>=4.66.0
numpy>=1.26.0
